fcmlib initial commit
This library is a variation of the gcmlib.

The only necessary change to make this library work
as a Firebase Cloud Messaging library is a change
in the google send endpoint for be the FCM endpoint.

Use of the three characters gcm or GCM have also been
changed to fcm and FCM in labels and error messages.